+++ src/SoftTissue.Core/Models/Viscoelasticity/QuasiLinear/FungModelResult.cs
@@ -4,8 +4,6 @@
     {
         public double StressWithIntegralDerivative { get; set; }
 
-        public double StressWithElasticResponseDerivative { get; set; }
-        
         public double StressWithReducedRelaxationFunctionDerivative { get; set; }
     }
 }