+++ src/SoftTissue.Core/Operations/QuasiLinearViscoelasticity/CalculateStress/FungModel/CalculateFungModelStressSentivityAnalysis.cs
@@ -192,7 +192,7 @@ namespace SoftTissue.Core.Operations.QuasiLinearViscoelasticity.CalculateStress.
                 Strain = strain,
                 ReducedRelaxationFunction = reducedRelaxationFunction,
                 ElasticResponse = elasticResponse,
-                StressWithElasticResponseDerivative = stressWithElasticResponseDerivative,
+                Stress = stressWithElasticResponseDerivative,
                 StressWithReducedRelaxationFunctionDerivative = stressWithReducedRelaxationFunctionDerivative,
                 StressWithIntegralDerivative = stressWithIntegrationDerivative
             };